v8-evaluation/pyproject.toml
[project]
name = "studybuddy-v8"
version = "8.0.0"
description = "StudyBuddy v8 - Evaluation Infrastructure"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # LangChain ecosystem (1.x as of late 2025)
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-community>=0.4.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-qdrant>=1.0.0",

    # LangGraph (1.x as of late 2025)
    "langgraph>=1.0.0",

    # Vector database
    "qdrant-client>=1.12.0",

    # Utilities
    "python-dotenv>=1.0.0",

    # Persistence (v6+)
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",  # PostgreSQL driver for Vercel Postgres

    # Evaluation Infrastructure (v8)
    "ragas>=0.2.0",
    "langsmith>=0.1.0",
    "datasets>=2.0.0",
    "pandas>=2.0.0",
    "rapidfuzz>=3.0.0",  # Required by RAGAS for string distance
    "nest-asyncio>=1.6.0",  # Required for RAGAS to work within FastAPI
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]
